separete drawer and panelButtons
